backend/requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
certifi==2026.1.4
click==8.1.8
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
fastapi==0.128.5
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
packaging==26.0
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
python-dotenv==1.2.1
PyJWT==2.10.1
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
